move Memory to be an interface
