+++ host-module/processor/src/test/resources/NoPackage.java
@@ -1,7 +1,7 @@
 import com.dylibso.chicory.experimental.hostmodule.annotations.HostModule;
 import com.dylibso.chicory.experimental.hostmodule.annotations.WasmExport;
-import com.dylibso.chicory.runtime.HostFunction;
 import com.dylibso.chicory.runtime.Memory;
+import com.dylibso.chicory.runtime.HostFunction;
 import com.dylibso.chicory.wasm.ChicoryException;
 
 @HostModule("nopackage")
