+++ host-module/processor/src/test/resources/Simple.java
@@ -6,8 +6,8 @@ import com.dylibso.chicory.experimental.hostmodule.annotations.Buffer;
 import com.dylibso.chicory.experimental.hostmodule.annotations.CString;
 import com.dylibso.chicory.experimental.hostmodule.annotations.HostModule;
 import com.dylibso.chicory.experimental.hostmodule.annotations.WasmExport;
-import com.dylibso.chicory.runtime.HostFunction;
 import com.dylibso.chicory.runtime.Memory;
+import com.dylibso.chicory.runtime.HostFunction;
 import com.dylibso.chicory.wasm.ChicoryException;
 import java.util.Random;
