+++ runtime/src/main/java/com/dylibso/chicory/runtime/Instance.java
@@ -663,7 +663,7 @@ public class Instance {
                 if (memories.memoryCount() > 0) {
                     var defaultLimits = memories.getMemory(0).limits();
                     memory =
-                            requireNonNullElse(memoryFactory, Memory::new)
+                            requireNonNullElse(memoryFactory, ByteBufferMemory::new)
                                     .apply(requireNonNullElse(memoryLimits, defaultLimits));
                 }
             } else {
